This fix has no effect on Windows (!).
Added the following lines to CMOOSThread::Start():

pthread_attr_t attr;
pthread_attr_init(&attr);
pthread_attr_setdetachstate(&attr, PTHREAD_CREATE_DETACHED);

...and changed the next line from:

int Status = pthread_create( &m_nThreadID,NULL,CallbackProc,this);

to

int Status = pthread_create( &m_nThreadID,&attr,CallbackProc,this);

The effect is that memory allocated for the thread on invocation of
pthread_create() is (automatically) returned to the system after
the thread finishes. On my system this was 8 MB per thread and if
one just starts and finishes some threads often enough the system
would sooner or later run out of memory.

On a side note: pthread_join() in CMOOSThread::Stop() seems to always
return with an error. As far as I understand, it should really not be
necessary to call it at all. I took the liberty to comment it out,
but I think it should/could be permanently removed.





git-svn-id: svn+ssh://login2/data/avl11/svnroot/MOOS/trunk@2259 5f621d73-a7ea-4b79-bc62-7e07601040e3