fix from mikerb
Hi Paul -

It seems there is a small problem when a process connects to the MOOSDB and registers for mail FOOBAR (when the last post to FOOBAR was long ago).

While it does indeed receive mail on FOOBAR, it is missing the source auxilliary information. Subsequent new mail on FOOBAR contains the auxilliary information, but it appears that MOOSDB does not hold onto this info.

It seems two additional book-keeping lines were missing in MOOSDB.cpp.
I'm attaching a revised MOOSDB.cpp with the two new lines denoted by "Added by mikerb". If you agree with the changes, would you consider including them in the latest version?

cheers,
Mike

Signed-off-by: Paul Newman <pnewman@robots.ox.ac.uk>