tentative fix for a bug which Alon saw which causes hell when less than sizeof(int) bytes are received in the first read on a packet
I made GetBytesRequired() return the number of bytes required to make m_nByteCount pass through sizeof(int) bytes always. So if 4 were asked for and only 1 returned then it asks for three and so on.  I also moved the logic of ::Fill around so it only figures out message length when it can *and* needs to which is ONLY when m_nByteCount is sizeof(int)