adding end to end detail timing audit
This adds a new functionality enabled at compile time that has MOOS Clients transmit (over multicast) timing informaiton and CPU load information. The idea is that the new utility “atm” for Audit The MOOS, is used to colate and log this information. Turn this option on with “-DENABLE_DETAILED_TIMING_AUDIT” or the -ENABLE_DETAILED_TIMING_AUDIT option in cmake