Remove whitespace padding on last field
There is no benefit to padding the last field with spaces.
Additionally this fixes a bug that is caused because nothing is output if the message's datatype is MOOS_NOT_SET which causes the ss<<std::left<<std::setw(20); to get applied to the ss<<"\n"; statement resulting in padding at the start of the next line of output.