This is a putative fix for a case where a suspended client can cause somethign of a catastrophe.
A half closed socket (suspended client) cause the server thread to block while waiting for a file write to complete. The apprioach here is to not change the SendPkt function which is old and used everywhere, but intead to fire  a SIGUSR1 signal at the potentially blocked thread. This thread does no allocation of its own  - so pthread_exit(0) whilst harsh in this case is a reasonable compromise.